# mat: Use untransposeExtract in Dense.Mul