Please take a look.

Alternative to #989 

There still needs more paths to be "fully compliant" but this at least simplifies the code.