@kortschak 2019-06-11
`master` _vs_ `multfastpath`
<details>
<table class='benchstat'>
<tr><th>name</th><th>old time/op</th><th>new time/op</th><th>delta</th>
<tr><td>MulDense100Half-8</td><td>197µs ± 6%</td><td>292µs ±111%</td><td>~   </td><td>(p=0.393 n=10+10)</td>
<tr><td>MulDense100Tenth-8</td><td>67.2µs ±12%</td><td>68.9µs ±31%</td><td>~   </td><td>(p=0.436 n=10+10)</td>
<tr><td>MulDense1000Half-8</td><td>72.2ms ± 5%</td><td>64.6ms ± 1%</td><td>-10.47%</td><td>(p=0.000 n=10+10)</td>
<tr><td>MulDense1000Tenth-8</td><td>22.5ms ± 5%</td><td>20.9ms ± 1%</td><td>-7.25%</td><td>(p=0.000 n=10+10)</td>
<tr><td>MulDense1000Hundredth-8</td><td>9.35ms ±11%</td><td>8.33ms ± 1%</td><td>-10.91%</td><td>(p=0.000 n=10+10)</td>
<tr><td>MulDense1000Thousandth-8</td><td>5.69ms ± 5%</td><td>5.39ms ± 0%</td><td>-5.26%</td><td>(p=0.004 n=10+10)</td>
<tr><td>PreMulDense100Half-8</td><td>171µs ±20%</td><td>157µs ±16%</td><td>~   </td><td>(p=0.243 n=10+9)</td>
<tr><td>PreMulDense100Tenth-8</td><td>55.6µs ±17%</td><td>55.0µs ± 2%</td><td>~   </td><td>(p=0.515 n=10+8)</td>
<tr><td>PreMulDense1000Half-8</td><td>69.9ms ±16%</td><td>67.3ms ± 1%</td><td>~   </td><td>(p=0.897 n=10+8)</td>
<tr><td>PreMulDense1000Tenth-8</td><td>20.7ms ± 3%</td><td>20.7ms ± 5%</td><td>~   </td><td>(p=0.878 n=8+8)</td>
<tr><td>PreMulDense1000Hundredth-8</td><td>7.95ms ±12%</td><td>7.90ms ± 3%</td><td>~   </td><td>(p=0.447 n=10+9)</td>
<tr><td>PreMulDense1000Thousandth-8</td><td>4.54ms ± 2%</td><td>4.60ms ± 2%</td><td>+1.34%</td><td>(p=0.006 n=8+9)</td>
<tr><td>MulWorkspaceDense100Half-8</td><td>244µs ± 6%</td><td>251µs ± 7%</td><td>~   </td><td>(p=0.165 n=10+10)</td>
<tr><td>MulWorkspaceDense100Tenth-8</td><td>232µs ±10%</td><td>247µs ± 5%</td><td>+6.41%</td><td>(p=0.023 n=10+10)</td>
<tr><td>MulWorkspaceDense1000Half-8</td><td>105ms ± 1%</td><td>102ms ± 5%</td><td>~   </td><td>(p=0.089 n=10+10)</td>
<tr><td>MulWorkspaceDense1000Tenth-8</td><td>109ms ± 4%</td><td>102ms ± 4%</td><td>-5.73%</td><td>(p=0.000 n=10+10)</td>
<tr><td>MulWorkspaceDense1000Hundredth-8</td><td>110ms ± 3%</td><td>99ms ± 1%</td><td>-9.77%</td><td>(p=0.000 n=10+10)</td>
<tr><td>MulWorkspaceDense1000Thousandth-8</td><td>6.77ms ±30%</td><td>5.53ms ±14%</td><td>-18.38%</td><td>(p=0.009 n=10+10)</td>
</table>

<table class='benchstat'>
<tr><th>name</th><th>old alloc/op</th><th>new alloc/op</th><th>delta</th>
<tr><td>MulDense100Half-8</td><td>82.2kB ± 0%</td><td>82.2kB ± 0%</td><td>~   </td><td>(p=0.059 n=9+8)</td>
<tr><td>MulDense100Tenth-8</td><td>82.2kB ± 0%</td><td>82.2kB ± 0%</td><td>~   </td><td>(p=1.000 n=10+10)</td>
<tr><td>MulDense1000Half-8</td><td>8.00MB ± 0%</td><td>8.00MB ± 0%</td><td>~   </td><td>(p=0.742 n=10+10)</td>
<tr><td>MulDense1000Tenth-8</td><td>8.00MB ± 0%</td><td>8.00MB ± 0%</td><td>~   </td><td>(p=0.457 n=10+10)</td>
<tr><td>MulDense1000Hundredth-8</td><td>8.00MB ± 0%</td><td>8.00MB ± 0%</td><td>-0.00%</td><td>(p=0.020 n=10+10)</td>
<tr><td>MulDense1000Thousandth-8</td><td>8.00MB ± 0%</td><td>8.00MB ± 0%</td><td>~   </td><td>(p=0.454 n=10+9)</td>
<tr><td>PreMulDense100Half-8</td><td>176B ± 0%</td><td>176B ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>PreMulDense100Tenth-8</td><td>176B ± 0%</td><td>176B ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>PreMulDense1000Half-8</td><td>656B ± 0%</td><td>662B ± 2%</td><td>+0.91%</td><td>(p=0.011 n=10+10)</td>
<tr><td>PreMulDense1000Tenth-8</td><td>658B ± 0%</td><td>659B ± 1%</td><td>~   </td><td>(p=0.433 n=10+10)</td>
<tr><td>PreMulDense1000Hundredth-8</td><td>657B ± 0%</td><td>656B ± 0%</td><td>~   </td><td>(p=0.380 n=10+8)</td>
<tr><td>PreMulDense1000Thousandth-8</td><td>658B ± 1%</td><td>658B ± 1%</td><td>~   </td><td>(p=0.887 n=9+10)</td>
<tr><td>MulWorkspaceDense100Half-8</td><td>243B ± 2%</td><td>244B ± 1%</td><td>~   </td><td>(p=0.174 n=9+9)</td>
<tr><td>MulWorkspaceDense100Tenth-8</td><td>235B ± 8%</td><td>242B ± 1%</td><td>~   </td><td>(p=0.504 n=10+8)</td>
<tr><td>MulWorkspaceDense1000Half-8</td><td>525kB ± 0%</td><td>488kB ±15%</td><td>~   </td><td>(p=0.080 n=10+10)</td>
<tr><td>MulWorkspaceDense1000Tenth-8</td><td>168kB ± 9%</td><td>165kB ±11%</td><td>~   </td><td>(p=0.305 n=9+10)</td>
<tr><td>MulWorkspaceDense1000Hundredth-8</td><td>84.6kB ± 0%</td><td>84.6kB ± 0%</td><td>~   </td><td>(p=0.306 n=9+10)</td>
<tr><td>MulWorkspaceDense1000Thousandth-8</td><td>43.7kB ±19%</td><td>39.2kB ±19%</td><td>-10.26%</td><td>(p=0.011 n=10+10)</td>
</table>

<table class='benchstat'>
<tr><th>name</th><th>old allocs/op</th><th>new allocs/op</th><th>delta</th>
<tr><td>MulDense100Half-8</td><td>4.00 ± 0%</td><td>4.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>MulDense100Tenth-8</td><td>4.00 ± 0%</td><td>4.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>MulDense1000Half-8</td><td>4.00 ± 0%</td><td>4.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>MulDense1000Tenth-8</td><td>4.00 ± 0%</td><td>4.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>MulDense1000Hundredth-8</td><td>4.00 ± 0%</td><td>4.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>MulDense1000Thousandth-8</td><td>4.00 ± 0%</td><td>4.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>PreMulDense100Half-8</td><td>2.00 ± 0%</td><td>2.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>PreMulDense100Tenth-8</td><td>2.00 ± 0%</td><td>2.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>PreMulDense1000Half-8</td><td>2.00 ± 0%</td><td>2.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>PreMulDense1000Tenth-8</td><td>2.00 ± 0%</td><td>2.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>PreMulDense1000Hundredth-8</td><td>2.00 ± 0%</td><td>2.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>PreMulDense1000Thousandth-8</td><td>2.00 ± 0%</td><td>2.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>MulWorkspaceDense100Half-8</td><td>4.00 ± 0%</td><td>4.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>MulWorkspaceDense100Tenth-8</td><td>4.00 ± 0%</td><td>4.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>MulWorkspaceDense1000Half-8</td><td>4.00 ± 0%</td><td>4.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>MulWorkspaceDense1000Tenth-8</td><td>4.00 ± 0%</td><td>4.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>MulWorkspaceDense1000Hundredth-8</td><td>4.00 ± 0%</td><td>4.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>MulWorkspaceDense1000Thousandth-8</td><td>4.00 ± 0%</td><td>4.00 ± 0%</td><td>~   </td><td>(all equal)</td>
</table>
</details>

`master` _vs_ `mat/generalise-vector-mul`
<details>
<table class='benchstat'>
<tr><th>name</th><th>old time/op</th><th>new time/op</th><th>delta</th>
<tr><td>MulDense100Half-8</td><td>197µs ± 6%</td><td>182µs ± 2%</td><td>-7.30%</td><td>(p=0.000 n=10+8)</td>
<tr><td>MulDense100Tenth-8</td><td>67.2µs ±12%</td><td>60.2µs ± 3%</td><td>-10.38%</td><td>(p=0.000 n=10+10)</td>
<tr><td>MulDense1000Half-8</td><td>72.2ms ± 5%</td><td>64.4ms ± 1%</td><td>-10.69%</td><td>(p=0.000 n=10+9)</td>
<tr><td>MulDense1000Tenth-8</td><td>22.5ms ± 5%</td><td>20.8ms ± 1%</td><td>-7.68%</td><td>(p=0.000 n=10+9)</td>
<tr><td>MulDense1000Hundredth-8</td><td>9.35ms ±11%</td><td>8.27ms ± 0%</td><td>-11.52%</td><td>(p=0.000 n=10+10)</td>
<tr><td>MulDense1000Thousandth-8</td><td>5.69ms ± 5%</td><td>5.34ms ± 1%</td><td>-6.12%</td><td>(p=0.000 n=10+10)</td>
<tr><td>PreMulDense100Half-8</td><td>171µs ±20%</td><td>151µs ±14%</td><td>-12.16%</td><td>(p=0.019 n=10+10)</td>
<tr><td>PreMulDense100Tenth-8</td><td>55.6µs ±17%</td><td>50.4µs ± 5%</td><td>-9.26%</td><td>(p=0.000 n=10+10)</td>
<tr><td>PreMulDense1000Half-8</td><td>69.9ms ±16%</td><td>63.3ms ± 0%</td><td>-9.36%</td><td>(p=0.000 n=10+9)</td>
<tr><td>PreMulDense1000Tenth-8</td><td>20.7ms ± 3%</td><td>19.9ms ± 1%</td><td>-3.81%</td><td>(p=0.000 n=8+8)</td>
<tr><td>PreMulDense1000Hundredth-8</td><td>7.95ms ±12%</td><td>8.30ms ± 0%</td><td>~   </td><td>(p=0.133 n=10+7)</td>
<tr><td>PreMulDense1000Thousandth-8</td><td>4.54ms ± 2%</td><td>4.75ms ± 1%</td><td>+4.72%</td><td>(p=0.000 n=8+10)</td>
<tr><td>MulWorkspaceDense100Half-8</td><td>244µs ± 6%</td><td>235µs ±15%</td><td>~   </td><td>(p=0.123 n=10+10)</td>
<tr><td>MulWorkspaceDense100Tenth-8</td><td>232µs ±10%</td><td>247µs ±17%</td><td>~   </td><td>(p=0.247 n=10+10)</td>
<tr><td>MulWorkspaceDense1000Half-8</td><td>105ms ± 1%</td><td>100ms ± 1%</td><td>-4.43%</td><td>(p=0.000 n=10+10)</td>
<tr><td>MulWorkspaceDense1000Tenth-8</td><td>109ms ± 4%</td><td>101ms ± 0%</td><td>-6.99%</td><td>(p=0.000 n=10+8)</td>
<tr><td>MulWorkspaceDense1000Hundredth-8</td><td>110ms ± 3%</td><td>100ms ± 0%</td><td>-8.88%</td><td>(p=0.000 n=10+10)</td>
<tr><td>MulWorkspaceDense1000Thousandth-8</td><td>6.77ms ±30%</td><td>5.93ms ±17%</td><td>-12.36%</td><td>(p=0.043 n=10+10)</td>
</table>

<table class='benchstat'>
<tr><th>name</th><th>old alloc/op</th><th>new alloc/op</th><th>delta</th>
<tr><td>MulDense100Half-8</td><td>82.2kB ± 0%</td><td>82.2kB ± 0%</td><td>~   </td><td>(p=0.440 n=9+10)</td>
<tr><td>MulDense100Tenth-8</td><td>82.2kB ± 0%</td><td>82.2kB ± 0%</td><td>~   </td><td>(p=0.065 n=10+9)</td>
<tr><td>MulDense1000Half-8</td><td>8.00MB ± 0%</td><td>8.00MB ± 0%</td><td>~   </td><td>(p=0.983 n=10+10)</td>
<tr><td>MulDense1000Tenth-8</td><td>8.00MB ± 0%</td><td>8.00MB ± 0%</td><td>~   </td><td>(p=0.620 n=10+9)</td>
<tr><td>MulDense1000Hundredth-8</td><td>8.00MB ± 0%</td><td>8.00MB ± 0%</td><td>-0.00%</td><td>(p=0.041 n=10+9)</td>
<tr><td>MulDense1000Thousandth-8</td><td>8.00MB ± 0%</td><td>8.00MB ± 0%</td><td>~   </td><td>(p=0.077 n=10+10)</td>
<tr><td>PreMulDense100Half-8</td><td>176B ± 0%</td><td>176B ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>PreMulDense100Tenth-8</td><td>176B ± 0%</td><td>176B ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>PreMulDense1000Half-8</td><td>656B ± 0%</td><td>661B ± 0%</td><td>+0.76%</td><td>(p=0.000 n=10+7)</td>
<tr><td>PreMulDense1000Tenth-8</td><td>658B ± 0%</td><td>658B ± 1%</td><td>~   </td><td>(p=0.611 n=10+9)</td>
<tr><td>PreMulDense1000Hundredth-8</td><td>657B ± 0%</td><td>656B ± 0%</td><td>~   </td><td>(p=0.448 n=10+9)</td>
<tr><td>PreMulDense1000Thousandth-8</td><td>658B ± 1%</td><td>658B ± 1%</td><td>~   </td><td>(p=0.790 n=9+10)</td>
<tr><td>MulWorkspaceDense100Half-8</td><td>243B ± 2%</td><td>241B ± 3%</td><td>~   </td><td>(p=0.205 n=9+9)</td>
<tr><td>MulWorkspaceDense100Tenth-8</td><td>235B ± 8%</td><td>244B ± 2%</td><td>~   </td><td>(p=0.120 n=10+8)</td>
<tr><td>MulWorkspaceDense1000Half-8</td><td>525kB ± 0%</td><td>467kB ± 0%</td><td>-11.10%</td><td>(p=0.000 n=10+9)</td>
<tr><td>MulWorkspaceDense1000Tenth-8</td><td>168kB ± 9%</td><td>169kB ± 2%</td><td>+0.76%</td><td>(p=0.046 n=9+9)</td>
<tr><td>MulWorkspaceDense1000Hundredth-8</td><td>84.6kB ± 0%</td><td>84.6kB ± 0%</td><td>-0.00%</td><td>(p=0.000 n=9+10)</td>
<tr><td>MulWorkspaceDense1000Thousandth-8</td><td>43.7kB ±19%</td><td>40.8kB ± 4%</td><td>-6.57%</td><td>(p=0.037 n=10+10)</td>
</table>

<table class='benchstat'>
<tr><th>name</th><th>old allocs/op</th><th>new allocs/op</th><th>delta</th>
<tr><td>MulDense100Half-8</td><td>4.00 ± 0%</td><td>4.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>MulDense100Tenth-8</td><td>4.00 ± 0%</td><td>4.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>MulDense1000Half-8</td><td>4.00 ± 0%</td><td>4.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>MulDense1000Tenth-8</td><td>4.00 ± 0%</td><td>4.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>MulDense1000Hundredth-8</td><td>4.00 ± 0%</td><td>4.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>MulDense1000Thousandth-8</td><td>4.00 ± 0%</td><td>4.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>PreMulDense100Half-8</td><td>2.00 ± 0%</td><td>2.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>PreMulDense100Tenth-8</td><td>2.00 ± 0%</td><td>2.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>PreMulDense1000Half-8</td><td>2.00 ± 0%</td><td>2.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>PreMulDense1000Tenth-8</td><td>2.00 ± 0%</td><td>2.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>PreMulDense1000Hundredth-8</td><td>2.00 ± 0%</td><td>2.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>PreMulDense1000Thousandth-8</td><td>2.00 ± 0%</td><td>2.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>MulWorkspaceDense100Half-8</td><td>4.00 ± 0%</td><td>4.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>MulWorkspaceDense100Tenth-8</td><td>4.00 ± 0%</td><td>4.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>MulWorkspaceDense1000Half-8</td><td>4.00 ± 0%</td><td>4.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>MulWorkspaceDense1000Tenth-8</td><td>4.00 ± 0%</td><td>4.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>MulWorkspaceDense1000Hundredth-8</td><td>4.00 ± 0%</td><td>4.00 ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>MulWorkspaceDense1000Thousandth-8</td><td>4.00 ± 0%</td><td>4.00 ± 0%</td><td>~   </td><td>(all equal)</td>
</table>
</details>

`mulfastpath` _vs_ `mat/generalise-vector-mul`
<details>
<table class='benchstat'>
<tr><th>name</th><th>old time/op</th><th>new time/op</th><th>delta</th>
<tr><td>MulDense100Half-8</td><td>292µs ±111%</td><td>182µs ± 2%</td><td>~   </td><td>(p=0.055 n=10+8)</td>
<tr><td>MulDense100Tenth-8</td><td>68.9µs ±31%</td><td>60.2µs ± 3%</td><td>-12.56%</td><td>(p=0.034 n=10+10)</td>
<tr><td>MulDense1000Half-8</td><td>64.6ms ± 1%</td><td>64.4ms ± 1%</td><td>~   </td><td>(p=0.113 n=10+9)</td>
<tr><td>MulDense1000Tenth-8</td><td>20.9ms ± 1%</td><td>20.8ms ± 1%</td><td>-0.46%</td><td>(p=0.013 n=10+9)</td>
<tr><td>MulDense1000Hundredth-8</td><td>8.33ms ± 1%</td><td>8.27ms ± 0%</td><td>-0.68%</td><td>(p=0.004 n=10+10)</td>
<tr><td>MulDense1000Thousandth-8</td><td>5.39ms ± 0%</td><td>5.34ms ± 1%</td><td>-0.91%</td><td>(p=0.000 n=10+10)</td>
<tr><td>PreMulDense100Half-8</td><td>157µs ±16%</td><td>151µs ±14%</td><td>~   </td><td>(p=0.211 n=9+10)</td>
<tr><td>PreMulDense100Tenth-8</td><td>55.0µs ± 2%</td><td>50.4µs ± 5%</td><td>-8.38%</td><td>(p=0.000 n=8+10)</td>
<tr><td>PreMulDense1000Half-8</td><td>67.3ms ± 1%</td><td>63.3ms ± 0%</td><td>-5.86%</td><td>(p=0.000 n=8+9)</td>
<tr><td>PreMulDense1000Tenth-8</td><td>20.7ms ± 5%</td><td>19.9ms ± 1%</td><td>-3.69%</td><td>(p=0.001 n=8+8)</td>
<tr><td>PreMulDense1000Hundredth-8</td><td>7.90ms ± 3%</td><td>8.30ms ± 0%</td><td>+5.05%</td><td>(p=0.000 n=9+7)</td>
<tr><td>PreMulDense1000Thousandth-8</td><td>4.60ms ± 2%</td><td>4.75ms ± 1%</td><td>+3.33%</td><td>(p=0.000 n=9+10)</td>
<tr><td>DenseMulTransDense100Half-8</td><td>269µs ± 3%</td><td>266µs ± 1%</td><td>-1.22%</td><td>(p=0.043 n=10+10)</td>
<tr><td>DenseMulTransDense100Tenth-8</td><td>271µs ± 1%</td><td>268µs ± 2%</td><td>-1.25%</td><td>(p=0.004 n=8+10)</td>
<tr><td>DenseMulTransDense1000Tenth-8</td><td>98.5ms ± 3%</td><td>95.9ms ± 0%</td><td>-2.69%</td><td>(p=0.002 n=10+9)</td>
<tr><td>DenseMulTransDense1000Hundredth-8</td><td>97.5ms ± 2%</td><td>95.8ms ± 0%</td><td>-1.76%</td><td>(p=0.035 n=9+10)</td>
<tr><td>DenseMulTransDense1000Thousandth-8</td><td>97.1ms ± 2%</td><td>95.7ms ± 0%</td><td>-1.38%</td><td>(p=0.028 n=9+10)</td>
<tr><td>DenseMulTransDenseSym100Half-8</td><td>265µs ± 3%</td><td>265µs ± 2%</td><td>~   </td><td>(p=0.912 n=10+10)</td>
<tr><td>DenseMulTransDenseSym100Tenth-8</td><td>268µs ± 2%</td><td>265µs ± 1%</td><td>~   </td><td>(p=0.165 n=10+10)</td>
<tr><td>DenseMulTransDenseSym1000Half-8</td><td>100ms ± 1%</td><td>100ms ± 0%</td><td>~   </td><td>(p=0.105 n=10+10)</td>
<tr><td>DenseMulTransDenseSym1000Tenth-8</td><td>101ms ± 2%</td><td>100ms ± 0%</td><td>-1.33%</td><td>(p=0.002 n=8+10)</td>
<tr><td>DenseMulTransDenseSym1000Hundredth-8</td><td>102ms ± 3%</td><td>100ms ± 0%</td><td>-2.28%</td><td>(p=0.000 n=10+9)</td>
<tr><td>DenseMulTransDenseSym1000Thousandth-8</td><td>108ms ±11%</td><td>100ms ± 0%</td><td>-7.18%</td><td>(p=0.000 n=10+9)</td>
<tr><td>MulWorkspaceDense100Half-8</td><td>251µs ± 7%</td><td>235µs ±15%</td><td>-6.67%</td><td>(p=0.023 n=10+10)</td>
<tr><td>MulWorkspaceDense100Tenth-8</td><td>247µs ± 5%</td><td>247µs ±17%</td><td>~   </td><td>(p=0.912 n=10+10)</td>
<tr><td>MulWorkspaceDense1000Half-8</td><td>102ms ± 5%</td><td>100ms ± 1%</td><td>~   </td><td>(p=0.481 n=10+10)</td>
<tr><td>MulWorkspaceDense1000Tenth-8</td><td>102ms ± 4%</td><td>101ms ± 0%</td><td>~   </td><td>(p=0.515 n=10+8)</td>
<tr><td>MulWorkspaceDense1000Hundredth-8</td><td>99.2ms ± 1%</td><td>100.1ms ± 0%</td><td>+0.98%</td><td>(p=0.000 n=10+10)</td>
<tr><td>MulWorkspaceDense1000Thousandth-8</td><td>5.53ms ±14%</td><td>5.93ms ±17%</td><td>~   </td><td>(p=0.143 n=10+10)</td>
<tr><td>TriMul/BenchmarkTriMul100-8</td><td>1.56ms ± 1%</td><td>1.55ms ± 1%</td><td>-1.07%</td><td>(p=0.002 n=10+8)</td>
<tr><td>TriMul/BenchmarkTriMul200-8</td><td>12.3ms ± 3%</td><td>12.4ms ± 8%</td><td>~   </td><td>(p=0.481 n=10+10)</td>
<tr><td>TriMul/BenchmarkTriMul400-8</td><td>105ms ± 3%</td><td>104ms ± 2%</td><td>~   </td><td>(p=0.211 n=10+9)</td>
<tr><td>TriMul/BenchmarkTriMul800-8</td><td>850ms ± 1%</td><td>873ms ± 1%</td><td>+2.77%</td><td>(p=0.000 n=10+10)</td>
<tr><td>TriMul/BenchmarkTriMul1600-8</td><td>9.79s ± 2%</td><td>9.37s ± 1%</td><td>-4.23%</td><td>(p=0.000 n=9+10)</td>
<tr><td>TriMulDiag/BenchmarkTriMulDiag100-8</td><td>60.6µs ± 3%</td><td>58.0µs ± 0%</td><td>-4.31%</td><td>(p=0.000 n=10+8)</td>
<tr><td>TriMulDiag/BenchmarkTriMulDiag200-8</td><td>237µs ± 1%</td><td>230µs ± 1%</td><td>-2.92%</td><td>(p=0.000 n=7+9)</td>
<tr><td>TriMulDiag/BenchmarkTriMulDiag400-8</td><td>951µs ± 2%</td><td>913µs ± 1%</td><td>-4.04%</td><td>(p=0.000 n=10+8)</td>
<tr><td>TriMulDiag/BenchmarkTriMulDiag800-8</td><td>3.86ms ± 4%</td><td>3.65ms ± 1%</td><td>-5.44%</td><td>(p=0.000 n=10+8)</td>
<tr><td>TriMulDiag/BenchmarkTriMulDiag1600-8</td><td>16.4ms ± 4%</td><td>14.8ms ± 3%</td><td>-9.29%</td><td>(p=0.000 n=10+10)</td>
<tr><td>TriMul2Diag/BenchmarkTriMul2Diag100-8</td><td>2.79µs ± 2%</td><td>2.88µs ± 1%</td><td>+3.18%</td><td>(p=0.000 n=9+10)</td>
<tr><td>TriMul2Diag/BenchmarkTriMul2Diag200-8</td><td>7.65µs ± 1%</td><td>7.37µs ± 1%</td><td>-3.74%</td><td>(p=0.000 n=9+10)</td>
<tr><td>TriMul2Diag/BenchmarkTriMul2Diag400-8</td><td>26.5µs ± 2%</td><td>25.7µs ± 3%</td><td>-2.98%</td><td>(p=0.000 n=10+10)</td>
<tr><td>TriMul2Diag/BenchmarkTriMul2Diag800-8</td><td>86.6µs ± 2%</td><td>84.8µs ± 2%</td><td>-2.03%</td><td>(p=0.029 n=10+10)</td>
<tr><td>TriMul2Diag/BenchmarkTriMul2Diag1600-8</td><td>781µs ± 1%</td><td>726µs ± 1%</td><td>-7.05%</td><td>(p=0.000 n=9+10)</td>
</table>

<table class='benchstat'>
<tr><th>name</th><th>old alloc/op</th><th>new alloc/op</th><th>delta</th>
<tr><td>MulDense100Half-8</td><td>82.2kB ± 0%</td><td>82.2kB ± 0%</td><td>~   </td><td>(p=0.294 n=8+10)</td>
<tr><td>MulDense100Tenth-8</td><td>82.2kB ± 0%</td><td>82.2kB ± 0%</td><td>~   </td><td>(p=0.173 n=10+9)</td>
<tr><td>MulDense1000Half-8</td><td>8.00MB ± 0%</td><td>8.00MB ± 0%</td><td>~   </td><td>(p=0.831 n=10+10)</td>
<tr><td>MulDense1000Tenth-8</td><td>8.00MB ± 0%</td><td>8.00MB ± 0%</td><td>+0.00%</td><td>(p=0.013 n=10+9)</td>
<tr><td>MulDense1000Hundredth-8</td><td>8.00MB ± 0%</td><td>8.00MB ± 0%</td><td>~   </td><td>(p=0.340 n=10+9)</td>
<tr><td>MulDense1000Thousandth-8</td><td>8.00MB ± 0%</td><td>8.00MB ± 0%</td><td>~   </td><td>(p=0.146 n=9+10)</td>
<tr><td>PreMulDense100Half-8</td><td>176B ± 0%</td><td>176B ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>PreMulDense100Tenth-8</td><td>176B ± 0%</td><td>176B ± 0%</td><td>~   </td><td>(all equal)</td>
<tr><td>PreMulDense1000Half-8</td><td>662B ± 2%</td><td>661B ± 0%</td><td>~   </td><td>(p=0.499 n=10+7)</td>
<tr><td>PreMulDense1000Tenth-8</td><td>659B ± 1%</td><td>658B ± 1%</td><td>~   </td><td>(p=0.674 n=10+9)</td>
<tr><td>PreMulDense1000Hundredth-8</td><td>656B ± 0%</td><td>656B ± 0%</td><td>~   </td><td>(p=0.823 n=8+9)</td>
…